Line 45, make sure variable m is an integer

+++ interpolation/splines/misc.py
@@ -42,7 +42,7 @@ def cartesian(arrays, out=None):
     if out is None:
         out = np.zeros([n, len(arrays)], dtype=dtype)
 
-    m = n / arrays[0].size
+    m = int( n / arrays[0].size)
     out[:,0] = np.repeat(arrays[0], m)
     if arrays[1:]:
         cartesian(arrays[1:], out=out[0:m,1:])
@@ -53,9 +53,9 @@ def cartesian(arrays, out=None):
 def mlinspace(a,b,orders,out=None):
 
     import numpy
-    
+
     sl = [numpy.linspace(a[i],b[i],orders[i]) for i in range(len(a))]
-    
+
     if out is None:
         out = cartesian(sl)
     else:
